Issue Fix 50
This closed ticket describes an launcher error if lsd slam viewer will
be launched in an roslaunch file https://github.com/tum-
vision/lsd_slam/issues/55.

I wrote an fix for this in the main_viewer.cpp how checks if an bag file
is containing in arguments and load the rosfileloop or otherwise an
rosThreadLoop. Now all roslaunch files should be working.